# api/openapi.yaml
openapi: "3.0.0"
info:
  title: CodeEdu API
  version: "1.0.0"
  description: Gamified programming education platform API

servers:
  - url: https://code-edu.cc.cd/api/v1
    description: Production server

paths:
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { type: object }
                  accessToken: { type: string }
                  refreshToken: { type: string }

  /auth/register:
    post:
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                email: { type: string }
                password: { type: string }
                displayName: { type: string }
      responses:
        "201":
          description: Registration successful

  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: Token refreshed

  /auth/me:
    get:
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user data

  /users/profile/{username}:
    get:
      summary: Get user profile
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User profile data

  /users/dashboard:
    get:
      summary: Get user dashboard
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard data

  /lessons:
    get:
      summary: List lessons
      parameters:
        - name: language
          in: query
          schema: { type: string, default: python }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: List of lessons

  /lessons/{id}:
    get:
      summary: Get lesson detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Lesson detail

  /lessons/{id}/submit:
    post:
      summary: Submit lesson answer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
                answer: { type: string }
      responses:
        "200":
          description: Submission result

  /leaderboard:
    get:
      summary: Get leaderboard
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [global, friends] }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: Leaderboard data

  /challenges:
    get:
      summary: List challenges
      parameters:
        - name: language
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of challenges

  /search:
    get:
      summary: Search platform
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Search results

  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service health status

  /admin/stats:
    get:
      summary: Get admin statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform statistics

  /admin/users:
    get:
      summary: List all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT